fix: keep pkg/factorysessions within package file-count limit
Move NewSessionResponseEventStore alias coverage into api_test.go instead of adding a sixteenth root package file.
